pyproject.toml
[project]
name = "macos-agent"
version = "0.1.0"
description = "Claude AI agent with native macOS computer control"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.41.0",
    "anthropic[bedrock,vertex]>=0.39.0",
    "jsonschema>=4.22.0",
    "boto3>=1.28.57",
    "google-auth>=2,<3",
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.7",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["computer_use_demo"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long
